+++ src/test/kotlin/LoaderNameTestConfig.kt
@@ -1,0 +1,11 @@
+import dev.s7a.ktconfig.KtConfig
+
+@KtConfig
+data class ServerConfig(
+    val nested: Nested,
+) {
+    @KtConfig(loaderName = "Custom{CLASS_NAME}Loader")
+    data class Nested(
+        val serverName: String,
+    )
+}